Move repository list to domain
