+++ src/domain/Repositories.ts
@@ -1,0 +1,10 @@
+export const Repositories = {
+    InstanceRepository: "instanceRepository",
+    StorageRepository: "storageRepository",
+    DownloadRepository: "downloadRepository",
+    GitHubRepository: "githubRepository",
+    AggregatedRepository: "aggregatedRepository",
+    EventsRepository: "eventsRepository",
+    MetadataRepository: "metadataRepository",
+    TransformationRepository: "transformationsRepository",
+};
