+++ src/domain/instance/usecases/ValidateInstanceUseCase.ts
@@ -1,8 +1,8 @@
 import i18n from "../../../locales";
-import { Repositories } from "../../../presentation/CompositionRoot";
 import { Either } from "../../common/entities/Either";
 import { UseCase } from "../../common/entities/UseCase";
 import { RepositoryFactory } from "../../common/factories/RepositoryFactory";
+import { Repositories } from "../../Repositories";
 import { Instance } from "../entities/Instance";
 import { InstanceRepositoryConstructor } from "../repositories/InstanceRepository";
