+++ src/domain/synchronization/usecases/GenericSyncUseCase.ts
@@ -3,7 +3,6 @@ import _ from "lodash";
 import i18n from "../../../locales";
 import SyncReport from "../../../models/syncReport";
 import SyncRule from "../../../models/syncRule";
-import { Repositories } from "../../../presentation/CompositionRoot";
 import { D2 } from "../../../types/d2";
 import { SynchronizationBuilder } from "../../../types/synchronization";
 import { cache } from "../../../utils/cache";
@@ -21,6 +20,7 @@ import { MetadataPackage } from "../../metadata/entities/MetadataEntities";
 import { MetadataRepositoryConstructor } from "../../metadata/repositories/MetadataRepository";
 import { DeletedMetadataSyncUseCase } from "../../metadata/usecases/DeletedMetadataSyncUseCase";
 import { MetadataSyncUseCase } from "../../metadata/usecases/MetadataSyncUseCase";
+import { Repositories } from "../../Repositories";
 import { Namespace } from "../../storage/Namespaces";
 import { StorageRepositoryConstructor } from "../../storage/repositories/StorageRepository";
 import { TransformationRepositoryConstructor } from "../../transformations/repositories/TransformationRepository";
